+++ test/test_counting_set.cpp
@@ -42,17 +42,17 @@ int main(int argc, char **argv) {
     cset.async_insert("apple");
     cset.async_insert("red");
 
-    ASSERT_RELEASE(cset.count("dog") == world.size());
-    ASSERT_RELEASE(cset.count("apple") == world.size());
-    ASSERT_RELEASE(cset.count("red") == world.size());
+    ASSERT_RELEASE(cset.count("dog") == (size_t)world.size());
+    ASSERT_RELEASE(cset.count("apple") == (size_t)world.size());
+    ASSERT_RELEASE(cset.count("red") == (size_t)world.size());
     ASSERT_RELEASE(cset.size() == 3);
 
     auto count_map = cset.all_gather({"dog", "cat", "apple"});
-    ASSERT_RELEASE(count_map["dog"] == world.size());
-    ASSERT_RELEASE(count_map["apple"] == world.size());
+    ASSERT_RELEASE(count_map["dog"] == (size_t)world.size());
+    ASSERT_RELEASE(count_map["apple"] == (size_t)world.size());
     ASSERT_RELEASE(cset.count("cat") == 0);
 
-    ASSERT_RELEASE(cset.count_all() == 3 * world.size());
+    ASSERT_RELEASE(cset.count_all() == 3 * (size_t)world.size());
   }
 
   return 0;
